--- 01-starting-project/src/components/TabButton.jsx
@@ -1,7 +1,0 @@
-export default function TabButton({ children, onSelect }) {
-	return (
-		<li>
-			<button onClick={onSelect}>{children}</button>
-		</li>
-	);
-}
