+++ 03-react-essentials/src/assets/react-core-concepts.png
