+++ 03-react-essentials/src/assets/state-mgmt.png
