+++ 03-react-essentials/src/components/CoreConcept.jsx
