+++ 03-react-essentials/src/components/TabButton.jsx
@@ -1,0 +1,9 @@
+export default function TabButton({ children, onSelect, isSelected }) {
+	return (
+		<li>
+			<button className={isSelected ? 'active' : undefined } onClick={onSelect}>
+				{children}
+			</button>
+		</li>
+	);
+}
