+++ 03-react-essentials/vite.config.js
